ocr_app/requirements.txt
Flask
pillow
numpy
pyocr
pytesseract
cnocr
opencv-python
pycorrector
flask_cors
onnxruntime
kenlm
deepl